プリセット詳細をコンポーネント分割、必要ファイルの追加
