+++ src/SavedPrompt.vue
@@ -23,60 +23,18 @@
                     </div>
                 </div>
             </section>
-            <section class="preset-detail">
-                <div v-if="selectedPreset !== null">
-                    <div class="title-area">
-                        <h2>{{ selectedPreset.description }}</h2>
-                        <a :href="'./register/commands.php?preset_id=' + selectedPreset.preset_id" class="btn-common blue">編集</a>
-                        <!-- <button class="btn-common blue">編集</button>
-                        <button class="btn-common green" :style="'display:none;'">保存</button> -->
-                    </div>
-                    <ul class="data-list">
-                        <li class="image">
-                            <img :src="selectedPreset.originalImage" alt="">
-                        </li>
-                        <li class="nsfw">
-                            <h3>nsfw</h3>
-                            <p>{{ selectedPreset.nsfw_display }}</p>
-                        </li>
-                        <li class="prompt copy">
-                            <h3>プロンプト</h3>
-                            <button 
-                                :class="[enhanceBraceMessage === '( )に変換' ? 'btn-common blue':'btn-common green']" 
-                                @click="toggleEnhanceBrace()"
-                            >{{ enhanceBraceMessage }}</button>
-                            <p @click="copyText(selectedPreset.commands, 'プロンプト')">{{ selectedPreset.commands }}</p>
-                        </li>
-                        <li class="prompt-ban copy">
-                            <h3>BANプロンプト</h3>
-                            <button 
-                                :class="[enhanceBraceMessage === '( )に変換' ? 'btn-common blue':'btn-common green']"  
-                                @click="toggleEnhanceBrace()"
-                            >{{ enhanceBraceMessage }}</button>
-                            <p @click="copyText(selectedPreset.commands_ban, 'BANプロンプト')">{{ selectedPreset.commands_ban }}</p>
-                        </li>
-                        <li class="seed copy">
-                            <h3>シード値</h3>
-                            <p @click="copyText(selectedPreset.seed, 'シード値')">{{ selectedPreset.seed }}</p>
-                        </li>
-                        <li class="resolution">
-                            <h3>解像度</h3>
-                            <p>{{ selectedPreset.resolution }}</p>
-                        </li>
-                        <li class="other">
-                            <h3>備考</h3>
-                            <p>{{ selectedPreset.others }}</p>
-                        </li>
-                    </ul>
-                </div>
-            </section>
+            <SelectedPresetComponent 
+                :selected="selectedPreset"
+                @setCopyAlertText="setCopyAlertText"
+            />
         </div>
     </div>
 </template>
 <script lang="ts">
 import fetchData from './assets/ts/fetchData'
 import HeaderComponent from './components/HeaderComponent.vue'
-import SearchBoxComponent from './components/SearchBoxComponent.vue'
+import SearchBoxComponent from './components/saver/SearchBoxComponent.vue'
+import SelectedPresetComponent from './components/saver/SelectedPresetComponent.vue'
 import './assets/scss/savedPrompt.scss'
 import { ref, onMounted } from 'vue'
 import axios from 'axios'
@@ -85,10 +43,12 @@ export default {
     components: {
         HeaderComponent,
         SearchBoxComponent,
+        SelectedPresetComponent,
     },
     setup() {
         // ログインユーザーの登録プリセット一覧
         const savedPromptList = ref<any>([])
+        console.log(JSON.stringify(savedPromptList.value))
         
         // 各プリセットに対応する画像とサムネイルのURLを取得
         const setImages = (presets: {[key: string]: any}[], currentPath: string) => {
@@ -169,37 +129,6 @@ export default {
                 })
         }
 
-        // 強化値の{}と()を切り替える
-        const enhanceBraceMessage = ref<string>('( )に変換')
-        const toggleEnhanceBrace = () => {
-            if (enhanceBraceMessage.value === '( )に変換') {
-                enhanceBraceMessage.value = '{ }に変換'
-                selectedPreset.value.commands = selectedPreset.value.commands.replaceAll(/\{/g, '(').replaceAll(/\}/g, ')')
-                selectedPreset.value.commands_ban = selectedPreset.value.commands_ban.replaceAll(/\{/g, '(').replaceAll(/\}/g, ')')
-            } else {
-                enhanceBraceMessage.value = '( )に変換'
-                selectedPreset.value.commands = selectedPreset.value.commands.replaceAll(/\(/g, '{').replaceAll(/\)/g, '}')
-                selectedPreset.value.commands_ban = selectedPreset.value.commands_ban.replaceAll(/\(/g, '{').replaceAll(/\)/g, '}')
-            }
-        }
-
-        // クリックした文字列をコピーする
-        const copyAlertText = ref<string>('')
-        const copyText = (text: string, name: string) => {
-            const href = location.href.substring(0,5)
-            if (href === 'https') {
-                navigator.clipboard.writeText(text)
-            } else if (href === 'http:') {
-                const input = document.createElement('input')
-                document.body.appendChild(input)
-                input.value = text
-                input.select()
-                document.execCommand('copy')
-                document.body.removeChild(input)
-            }
-            copyAlertText.value = selectedPreset.value.description + 'の' + name + 'をコピーしました。'
-        }
-
         // ログインユーザーIDを取得
         const user_id = ref<string>('')
         const getUserInfo = async() => {
@@ -208,6 +137,10 @@ export default {
                 .then(response => user_id.value = response.data.user_id)
                 .catch(error => console.log(error))
         }
+
+        // コピーした際のアラートを設定
+        const copyAlertText = ref<string>('')
+        const setCopyAlertText = (text: string) => copyAlertText.value = text
 
         // 画面ロード時、APIからログインユーザーの登録プロンプト一覧を取得
         onMounted(() => {
@@ -220,13 +153,12 @@ export default {
             selectedPreset,
             selectedPresetIndex,
             searchData: searchData,
-            enhanceBraceMessage: enhanceBraceMessage,
             copyAlertText,
             user_id,
+
             selectPreset,
             getPresetData,
-            toggleEnhanceBrace,
-            copyText,
+            setCopyAlertText,
         }
     }
 }
