+++ src/components/saver/ManagePromptComponent.vue
@@ -1,0 +1,22 @@
+<template lang="">
+    <div>
+        
+    </div>
+</template>
+<script lang="ts">
+import { ref } from 'vue'
+import '../../assets/scss/savedPrompt.scss'
+
+export default {
+    props: {
+
+    },
+    emits: [],
+    setup(props:any, context:any) {
+
+        return {
+            
+        }
+    }
+}
+</script>
